Magic 8 Ball Game
I made the graphics myself in Adobe InDesign, as shown in the PNGs, they are used to run a basic Magic Eight Ball Game. You can click in the center, type your question, and then click ask a question and you will get an answerback. I would like to fix the button's click zones as they are way too large but for now, this is pretty close to what I wanted to build.